# openapi/openapi.yaml
openapi: 3.0.4
info:
  title: Marketplace API
  description: |
    Products and orders for a small marketplace.
    Order lists use cursor pagination; order creation requires an Idempotency-Key.
    All errors are returned as application/problem+json (RFC 9457).
  version: 1.0.0

servers:
  - url: http://localhost:3000
    description: Local development server

security: []

paths:
  /products/{id}:
    get:
      operationId: getProductById
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the product to retrieve.
          schema:
            type: integer
      responses:
        '200':
          description: The requested product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

  /products:
    get:
      operationId: listProducts
      summary: List all products
      responses:
        '200':
          description: A list of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '400':
          description: The request includes an unknown query parameter
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

  /orders/{id}:
    get:
      operationId: getOrderById
      summary: Get an order by ID
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the order to retrieve.
          schema:
            type: integer
      responses:
        '200':
          description: The requested order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

  /orders:
    get:
      operationId: listOrders
      summary: List orders
      parameters:
        - name: limit
          in: query
          required: false
          description: Maximum number of orders to return per page.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: cursor
          in: query
          required: false
          description: >
            An opaque pagination token. Send back exactly the value received
            in the previous response's next_cursor to fetch the next page;
            never parse or construct this value yourself.
          schema:
            type: string
      responses:
        '200':
          description: A page of orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPage'
        '400':
          description: The cursor query parameter is malformed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    post:
      operationId: createOrder
      summary: Create an order
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          description: >
            A client-generated v4 UUID that uniquely identifies this order
            creation attempt. Retrying the request with the same key and the
            same request body returns the original response instead of
            creating a second order.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrder'
      responses:
        '201':
          description: The created order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: The request body is invalid, or the Idempotency-Key header is missing
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: The Idempotency-Key was already used with a different request body
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

components:
  schemas:
    Problem:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
      required:
        - type
        - title
        - status
        - detail
        - instance

    Product:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        price_cents:
          type: integer
        in_stock:
          type: boolean
      required:
        - id
        - title
        - price_cents
        - in_stock

    OrderItem:
      type: object
      additionalProperties: false
      properties:
        product_id:
          type: integer
        quantity:
          type: integer
          minimum: 1
      required:
        - product_id
        - quantity

    Order:
      type: object
      properties:
        id:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
          minItems: 1
        total_cents:
          type: integer
        status:
          type: string
          enum:
            - new
            - paid
            - shipped
        created_at:
          type: string
          format: date-time
      required:
        - id
        - items
        - total_cents
        - status
        - created_at

    CreateOrder:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
          minItems: 1
      required:
        - items

    OrderPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        next_cursor:
          type: string
          nullable: true
      required:
        - items
        - next_cursor